Calculate required bond for the given satoshi lot size in TBTCSystem
Instead of having the same code in two places - TBTCSystem and
DepositFunding, we now calculate the required bond only in TBTCSystem.
Instead of passing the bond value to requestNewKeep function,
DepositFunding passes the lot size and TBTCSystem does all the
calculations.

+++ solidity/contracts/test/system/TBTCSystemStub.sol
@@ -27,7 +27,7 @@ contract TBTCSystemStub is TBTCSystem {
         keepAddress = _keepAddress;
     }
 
-    function requestNewKeep(uint256, uint256) external payable returns (address _keepAddress) {
+    function requestNewKeep(uint64, uint256) external payable returns (address _keepAddress) {
         return keepAddress;
     }
 }
